Adding test for Specter deletion, adding specter deletion methods to facade and service, fixing specter and GI delete methods in DAO

+++ haunted-houses/haunted-houses-persistence/src/main/java/cz/muni/fi/pa165/hauntedhouses/dao/impl/GameInstanceDaoImpl.java
@@ -65,16 +65,14 @@ public class GameInstanceDaoImpl implements GameInstanceDao {
 
     @Override
     public void deleteGameInstance(GameInstance gameInstance) {
-        System.err.println("Deleting GI through DAO");
-        //em.remove(em.contains(gameInstance) ? gameInstance : em.merge(gameInstance));
-        /*if (em.contains(gameInstance)) {
+        if (em.contains(gameInstance)) {
+            gameInstance.getPlayer().setGameInstance(null);
             em.remove(gameInstance);
         } else {
             GameInstance merge = em.merge(gameInstance);
+            gameInstance.getPlayer().setGameInstance(null);
             em.remove(merge);
         }
-        em.flush();*/
-        em.remove(em.merge(gameInstance));
     }
 }
